[FIRST] Initializing openCL/openGL interoperability shared buffer (openCL 3.0), to all be refactored in more readable code, dynamic release to do

+++ includes/camera.hpp
@@ -6,7 +6,7 @@
 /*   By: tmoragli <tmoragli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/01 20:26:25 by tmoragli          #+#    #+#             */
-/*   Updated: 2024/10/12 01:05:19 by tmoragli         ###   ########.fr       */
+/*   Updated: 2024/10/13 01:39:27 by tmoragli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -19,13 +19,13 @@ namespace psys {
 		public:
 			camera() : position{0.0, 0.0, 0.0} {};
 			~camera() {};
-			void move(double forward, double strafe);
+			void move(float forward, float strafe);
 			void reset();
 
 			vec3 position;
-			double xangle = 0.0;
-			double yangle = 0.0;
-			double rotationspeed = 1;
-			double movementspeed = 0.1;
+			float xangle = 0.0;
+			float yangle = 0.0;
+			float rotationspeed = 1;
+			float movementspeed = 0.1;
 	};
 };
