+++ includes/matrix.hpp
@@ -6,7 +6,7 @@
 /*   By: tmoragli <tmoragli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/09/24 00:32:35 by tmoragli          #+#    #+#             */
-/*   Updated: 2024/10/12 01:05:19 by tmoragli         ###   ########.fr       */
+/*   Updated: 2024/10/13 01:39:57 by tmoragli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,15 +16,15 @@
 namespace psys {
 	class mat4 {
 		public:
-			std::array<std::array<double, 4>, 4> data;
+			std::array<std::array<float, 4>, 4> data;
 			mat4();
-			mat4(const std::array<std::array<double, 4>, 4> &cpy);
+			mat4(const std::array<std::array<float, 4>, 4> &cpy);
 
 			// Set as identity matrix
 			static mat4 identity();
-			static mat4 translate(double x, double y, double z);
-			static mat4 rotate(double angle, double x, double y, double z);
-			static mat4 perspective(double fov, double aspect, double near, double far);
+			static mat4 translate(float x, float y, float z);
+			static mat4 rotate(float angle, float x, float y, float z);
+			static mat4 perspective(float fov, float aspect, float near, float far);
 
 			mat4 operator*(const mat4 &other) const;
 			mat4 &operator*=(const mat4 &other);
