+++ srcs/camera.cpp
@@ -6,7 +6,7 @@
 /*   By: tmoragli <tmoragli@student.42.fr>          +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/10/01 20:27:33 by tmoragli          #+#    #+#             */
-/*   Updated: 2024/10/12 01:04:39 by tmoragli         ###   ########.fr       */
+/*   Updated: 2024/10/13 01:39:31 by tmoragli         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,21 +14,21 @@
 
 namespace psys
 {
-	void camera::move(double forward, double strafe)
+	void camera::move(float  forward, float  strafe)
 	{
 		// Calculate the direction based on the current angles
-		double radiansY = yangle * (M_PI / 180.0);
-		double radiansX = xangle * (M_PI / 180.0);
+		float radiansY = yangle * (M_PI / 180.0);
+		float radiansX = xangle * (M_PI / 180.0);
 
-		double scaleForward = forward * cos(radiansY);
+		float scaleForward = forward * cos(radiansY);
 
 		// Determine the forward movement vector
-		double forwardX = cos(radiansX) * scaleForward;
-		double forwardZ = sin(radiansX) * scaleForward;
+		float forwardX = cos(radiansX) * scaleForward;
+		float forwardZ = sin(radiansX) * scaleForward;
 
 		// Determine the strafe movement vector (perpendicular to forward)
-		double strafeX = cos(radiansX + M_PI / 2) * strafe;
-		double strafeZ = sin(radiansX + M_PI / 2) * strafe;
+		float strafeX = cos(radiansX + M_PI / 2) * strafe;
+		float strafeZ = sin(radiansX + M_PI / 2) * strafe;
 
 		// Update the camera position
 		position.z += (forwardX + strafeX) * movementspeed;
